Fix MSVC warning C4244 '=': conversion from '__int64' to 'int', possible loss of data
Use explicit cast and intptr_t (cf. https://github.com/coin-or/CoinUtils/commit/eeaadb027bb8118aee5117152a7383dd8141b23c)

+++ src/coin/CoinUtils/CoinOslFactorization2.cpp
@@ -501,8 +501,8 @@ static int c_ekkbtj4p_dense(const int nrow,const double * COIN_RESTRICT dluval,
     dv2=densew[0];
     for (k=0;k<nincol;k++) {
 #ifdef DEBUG
-      int kk=dlu1-dluval;
-      int jj = (densew+(nincol-k+1))-dwork1;
+      intptr_t kk=dlu1-dluval;
+      intptr_t jj = (densew+(nincol-k+1))-dwork1;
       int ll=hrowi[k+kk];
       if (ll!=jj) abort();
 #endif
